Split mochitest/head.js in head.js and helpers.js
