+++ src/test/mochitest/browser.ini
@@ -4,6 +4,7 @@ subsuite = devtools
 skip-if = (os == 'linux' && debug && bits == 32)
 support-files =
   head.js
+  helpers.js
   !/devtools/client/commandline/test/helpers.js
   !/devtools/client/shared/test/shared-head.js
   examples/babel/polyfill-bundle.js
